move coulomb input class to seperate module
