+++ runzi/runners/__init__.py
@@ -1,6 +1,7 @@
+from .inversion_inputs import CoulombRuptureSetsInput
 from .average_solutions import run_average_solutions
 from .build_manual_index import build_manual_index
-from .coulomb_rupture_sets import CoulombRuptureSetsInput, run_coulomb_rupture_sets
+from .coulomb_rupture_sets import run_coulomb_rupture_sets
 from .hazard_inputs import DisaggInput, HazardInput
 from .inversion import run_inversion
 from .inversion_diagnostics import run_inversion_diagnostics
