+++ test/runners/test_runner_inputs.py
@@ -6,11 +6,11 @@ from pydantic import ValidationError
 
 from runzi.automation.scaling.toshi_api.general_task import ModelType
 from runzi.runners import (
-    CoulombRuptureSetsInput,
     ScaleSolutionsInput,
     SubductionRuptureSetsInput,
     TimeDependentSolutionInput,
 )
+from runzi.runners.inversion_inputs import CoulombRuptureSetsInput
 from runzi.runners.runner_inputs import AverageSolutionsInput
 
 fixtures_path = Path(__file__).parent.parent / 'fixtures/runners'
